+++ src/limbo/limbo.hpp
@@ -2,22 +2,22 @@
 #define LIMBO_HPP
 
 #include <limbo/tools/macros.hpp>
-#include <limbo/stop/stopping_criteria.hpp>
-#include <limbo/stat/stats.hpp>
+#include <limbo/stopping_criteria.hpp>
+#include <limbo/stats.hpp>
 #include <limbo/tools/sys.hpp>
 #include <limbo/tools/rand.hpp>
 #include <limbo/bayes_opt/boptimizer.hpp>
 #include <limbo/bayes_opt/ehvi.hpp>
 #include <limbo/bayes_opt/nsbo.hpp>
 #include <limbo/bayes_opt/parego.hpp>
-#include <limbo/kernel/kernel_functions.hpp>
-#include <limbo/acqui/acquisition_functions.hpp>
-#include <limbo/mean/mean_functions.hpp>
-#include <limbo/inner_opt/inner_optimization.hpp>
+#include <limbo/kernel_functions.hpp>
+#include <limbo/acquisition_functions.hpp>
+#include <limbo/mean_functions.hpp>
+#include <limbo/inner_optimization.hpp>
 #include <limbo/model/gp.hpp>
 #include <limbo/model/gp_auto.hpp>
 #include <limbo/model/gp_auto_mean.hpp>
-#include <limbo/init/initialization_functions.hpp>
+#include <limbo/initialization_functions.hpp>
 #include <limbo/tools/parallel.hpp>
 #include <limbo/opt/rprop.hpp>
