+++ src/limbo/stopping_criteria.hpp
